+++ satella/coding/concurrent/futures/__init__.py
@@ -1,4 +1,6 @@
 from .futures import Future, WrappingFuture, InvalidStateError
 from .call_in_future import call_in_future
+from .wrapped_executor import ExecutorWrapper
 
-__all__ = ['Future', 'WrappingFuture', 'InvalidStateError', 'call_in_future']
+__all__ = ['Future', 'WrappingFuture', 'InvalidStateError', 'call_in_future',
+           'ExecutorWrapper']