+++ CHANGELOG.md
@@ -1,1 +1,4 @@
 # v2.12.7
+
+* `call_in_future` callable will now return the Future
+* added `ExecutorWrapper`